implement autorization using Gates

+++ app/Http/Controllers/QuestionController.php
@@ -69,6 +69,12 @@ class QuestionController extends Controller
      */
     public function edit(Question $question)
     {
+
+        if(\Gate::denies('update-question',$question)){
+            
+            abort(403,'Access denied');
+        }
+
         return view('questions.edit',compact('question'));
     }
 
@@ -81,6 +87,10 @@ class QuestionController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if(\Gate::denies('update-question',$question)){
+            
+            abort(403,'Access denied');
+        }
         $question->update($request->only('title','body'));
         return redirect()->route('questions.index')->with('success','Your Question has been update.');
     }
@@ -93,6 +103,10 @@ class QuestionController extends Controller
      */
     public function destroy(Question $question)
     {
+        if(\Gate::denies('delete-question',$question)){
+            
+            abort(403,'Access denied');
+        }
         $question->delete();
         return redirect()->route('questions.index')->with('success','Your question has been deleted');
     }
